// packages/click-ui/src/components/Grid/types.ts
interface CellHeaderProps extends CellCommonProps {
  rowIndex?: never;
  type: 'header-cell';
  width: number;
}

interface CellBodyProps extends CellCommonProps {
  rowIndex: number;
  type: 'row-cell';
  width: number;
}

type CellComponentProps = CellHeaderProps | CellBodyProps;

export type CellProps = ComponentType<CellComponentProps>;
// ...
export type SelectionType = 'default' | 'selectIndirect' | 'selectDirect';
// ...
type IsAllSelectedType = {
  type: 'all';
};

type IsRowSelectedType = {
  type: 'row';
  row: number;
};

type IsColumnSelectedType = {
  type: 'column';
  column: number;
};

type IsCellSelectedType = {
  type: 'cell';
  row: number;
  column: number;
};

export type IsSelectedType =
  | IsAllSelectedType
  | IsRowSelectedType
  | IsColumnSelectedType
  | IsCellSelectedType;

export type SelectionTypeFn = (props: IsSelectedType) => SelectionType;

export type ColumnResizeFn = (
  columnIndex: number,
  newWidth: number,
  type: 'auto' | 'manual'
) => void;
// ...
export type ResizerPosition = {
  left: string;
  top?: string;
};

export type GetResizerPositionFn = (
  clientX: number,
  width: number,
  columnIndex: number
) => ResizerPosition;